// react-module/src/openapi/models/contractor-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ContractorData
 */
export interface ContractorData {
    /**
     * 
     * @type {Gid}
     * @memberof ContractorData
     */
    gid?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    acronym?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    name1?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    name2?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    name3?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    postCode?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    city?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    street?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    nip?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorData
     */
    personalIdNumber?: any;
}