// react-module/src/openapi/models/erp-order-with-ctx.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ErpOrderWithCtx
 */
export interface ErpOrderWithCtx {
    /**
     * 
     * @type {LocalErpOrderDto}
     * @memberof ErpOrderWithCtx
     */
    localErpOrder?: any;
    /**
     * 
     * @type {ErpOrder}
     * @memberof ErpOrderWithCtx
     */
    erpOrder?: any;
    /**
     * 
     * @type {number}
     * @memberof ErpOrderWithCtx
     */
    progress?: any;
}