// react-module/src/openapi/models/gid.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Gid
 */
export interface Gid {
    /**
     * 
     * @type {number}
     * @memberof Gid
     */
    type?: any;
    /**
     * 
     * @type {number}
     * @memberof Gid
     */
    company?: any;
    /**
     * 
     * @type {number}
     * @memberof Gid
     */
    number?: any;
    /**
     * 
     * @type {number}
     * @memberof Gid
     */
    counter?: any;
}