// react-module/src/openapi/models/local-erp-order-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface LocalErpOrderDto
 */
export interface LocalErpOrderDto {
    /**
     * 
     * @type {number}
     * @memberof LocalErpOrderDto
     */
    id?: any;
    /**
     * 
     * @type {GidDto}
     * @memberof LocalErpOrderDto
     */
    gid?: any;
    /**
     * 
     * @type {string}
     * @memberof LocalErpOrderDto
     */
    orderNumber?: any;
    /**
     * 
     * @type {string}
     * @memberof LocalErpOrderDto
     */
    description?: any;
    /**
     * 
     * @type {string}
     * @memberof LocalErpOrderDto
     */
    foreignDocument?: any;
    /**
     * 
     * @type {OrderWithElements}
     * @memberof LocalErpOrderDto
     */
    srcOrder?: any;
    /**
     * 
     * @type {Array&lt;WithoutErpOrder&gt;}
     * @memberof LocalErpOrderDto
     */
    productionOrders?: any;
    /**
     * 
     * @type {Date}
     * @memberof LocalErpOrderDto
     */
    deadline?: any;
    /**
     * 
     * @type {ContractorDto}
     * @memberof LocalErpOrderDto
     */
    contractor?: any;
}