// react-module/src/openapi/models/paged-result-erp-order-with-ctx.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PagedResultErpOrderWithCtx
 */
export interface PagedResultErpOrderWithCtx {
    /**
     * 
     * @type {Array&lt;ErpOrderWithCtx&gt;}
     * @memberof PagedResultErpOrderWithCtx
     */
    result?: any;
    /**
     * 
     * @type {number}
     * @memberof PagedResultErpOrderWithCtx
     */
    totalResults?: any;
}