// react-module/src/openapi/models/user-definition.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UserDefinition
 */
export interface UserDefinition {
    /**
     * 
     * @type {Gid}
     * @memberof UserDefinition
     */
    gid?: any;
    /**
     * 
     * @type {string}
     * @memberof UserDefinition
     */
    identificator?: any;
    /**
     * 
     * @type {string}
     * @memberof UserDefinition
     */
    name?: any;
    /**
     * 
     * @type {boolean}
     * @memberof UserDefinition
     */
    disabled?: any;
    /**
     * 
     * @type {number}
     * @memberof UserDefinition
     */
    frsId?: any;
    /**
     * 
     * @type {string}
     * @memberof UserDefinition
     */
    frsName?: any;
    /**
     * 
     * @type {string}
     * @memberof UserDefinition
     */
    frsDescription?: any;
}